move pyramidio files to separate package
